Fix tokensEarned type from int to double for fractional token splits
Backend now returns fractional tokensEarned (e.g. 4.5) after removing
Math.floor. Updated entity, model, datasource, state, and history sheet
to use double. Hide token breakdown card during CF submission phase.

+++ lib/domain/entities/engagement.dart
@@ -20,7 +20,7 @@ class Engagement with _$Engagement {
     required int requiredDurationSeconds,
     required List<SurveyResponse> answers,
     EngagementEvidence? evidence,
-    int? tokensEarned,
+    double? tokensEarned,
     String? failureReason,
     required int attemptNumber,
     required DateTime createdAt,
